requirements.txt
discord.py
requests~=2.28.2
numpy==1.24.3
matplotlib~=3.7.1
aiohttp~=3.8.4
pycountry~=22.3.5
pytz~=2023.3
colorlog~=6.7.0